Initial import of p5-Parse-CPAN-Packages version 2.31nb1 in the NetBSD Packages Collection.
The Comprehensive Perl Archive Network (CPAN) is a very useful
collection of Perl code. It has several indices of the files that
it hosts, including a file named "02packages.details.txt.gz" in the
"modules" directory. This file contains lots of useful information
and the Perl 5 module Parse::CPAN::Packages provides a simple
interface to the data contained within.